requirements.txt
flask==2.3.2
flask-cors==3.0.10
appium-python-client==2.9.0
Appium-Flutter-Finder==0.2.1
selenium==4.9.0
pytest==7.3.1
requests==2.28.2
python-dotenv==1.0.0 